pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentic-profile-auth"
version = "0.1.0"
description = "Agentic Profile Authentication Library"
authors = [
    { name = "Mike Prince" },
    { name = "Isabella Farley", email = "izzy@fondutech.com" }
]
dependencies = [
    "cryptography>=42.0.0",
    "pydantic>=2.0.0",
    "python-jose>=3.3.0",
    "loguru>=0.7.0",
    "aiohttp>=3.9.0",
    "did-resolver>=0.1.0"
]
requires-python = ">=3.9"
readme = "README.md"
license = { text = "MIT" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Security :: Cryptography",
    "Topic :: Software Development :: Libraries :: Python Modules"
]

[project.urls]
Homepage = "https://github.com/yourusername/agentic-profile-auth"
Repository = "https://github.com/yourusername/agentic-profile-auth.git"
Documentation = "https://agentic-profile-auth.readthedocs.io/"

[tool.hatch.build.targets.wheel]
packages = ["src/agentic_profile_auth"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.10.0"
]

[tool.mypy]
python_version = "3.8"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
line-length = 100
target-version = "py38" 